Type safety, cast cleanups, etc.
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@1401 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/command.cc
@@ -215,7 +215,7 @@ static void adjust_spells(void)
         return;
     }
 
-    int input_1 = (int) keyin;
+    int input_1 = keyin;
 
     if (!isalpha( input_1 ))
     {
@@ -265,7 +265,7 @@ static void adjust_spells(void)
         return;
     }
 
-    int input_2 = (int) keyin;
+    int input_2 = keyin;
 
     if (!isalpha( input_2 ))
     {
@@ -335,7 +335,7 @@ static void adjust_ability(void)
         return;
     }
 
-    int input_1 = (int) keyin;
+    int input_1 = keyin;
 
     if (!isalpha( input_1 ))
     {
@@ -384,7 +384,7 @@ static void adjust_ability(void)
         return;
     }
 
-    int input_2 = (int) keyin;
+    int input_2 = keyin;
 
     if (!isalpha( input_2 ))
     {
